requirements.txt
streamlit
openai-whisper
ffmpeg-python
torchaudio
